# openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Fitness360 API",
    "version": "1.0.0",
    "description": "Generated from App Router route files + agent tool registry (GYM-OS-C-006)"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "OAuth"
      }
    }
  },
  "paths": {
...
    "/challenges/participants": {
      "post": {
        "summary": "POST /challenges/participants",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      },
      "patch": {
        "summary": "PATCH /challenges/participants",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      },
      "delete": {
        "summary": "DELETE /challenges/participants",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
  }
}